+++ my-app/src/app.html
@@ -9,12 +9,14 @@
 	<body
 		data-sveltekit-preload-data="hover"
 		style="
-			width: 100vw;
+			width: 100%;
 			height: 100%;
 			padding: 0;
 			margin: 0;
-			font-family: fantasy;
+			font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif;
 			background-color: #d9d9d9;
+			overflow-y:scroll;
+			overflow-x: hidden;
 		"
 	>
 		<div style="display: contents">%sveltekit.body%</div>
