+++ my-app/src/lib/First.svelte
@@ -1,5 +1,35 @@
 <script>
+	import { onMount } from 'svelte';
 	export let character;
+	/**
+	 * @type {any[]}
+	 */
+	let backgrounds = []; // Array to hold the fetched backgrounds
+	/**
+	 * @type {{ desc: any; benefits: any; } | null}
+	 */
+	let selectedBackground = null; // To hold the currently selected background
+
+	onMount(async () => {
+		try {
+			const response = await fetch('https://api.open5e.com/v2/backgrounds/');
+			if (!response.ok) throw new Error('Failed to fetch backgrounds');
+			const data = await response.json();
+			backgrounds = data.results; // Assuming the API returns an object with a 'results' array
+			// Set the default selected background if any
+			if (backgrounds.length > 0) {
+				selectedBackground = backgrounds[0]; // Set the first background as default
+			}
+		} catch (error) {
+			console.error('Error fetching backgrounds:', error);
+		}
+	});
+
+	const onBackgroundChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		const selected = backgrounds.find(bg => bg.name === event.target.value);
+		$character.Background = selected ? selected.name : ''; // Update the character background
+		selectedBackground = selected; // Update selected background details
+	};
 	const onLevelChange = (/** @type {{ target: { value: any; }; }} */ event) => {
 		$character.Level = event.target.value;
 	};
@@ -8,9 +38,6 @@
 	};
 	const onClassChange = (/** @type {{ target: { value: any; }; }} */ event) => {
 		$character.Class = event.target.value;
-	};
-	const onBackgroundChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Background = event.target.value;
 	};
 	const onHitPointChange = (/** @type {{ target: { value: any; }; }} */ event) => {
 		$character.RolledHitpoints = event.target.value;
@@ -27,7 +54,7 @@
 			bind:value={$character.Level}
 			name="Level"
 			id="Level"
-			style="font-family: fantasy; font-size: 20px;"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
 		>
 			<option class="option" value={1}>1</option>
 			<option class="option" value={2}>2</option>
@@ -57,7 +84,7 @@
 			bind:value={$character.Race}
 			name="Race"
 			id="Race"
-			style="font-family: fantasy; font-size: 20px;"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
 		>
 			<option value="Dragonborn" class="option">Dragonborn</option>
 			<option value="Dwarf" class="option">Dwarf</option>
@@ -76,7 +103,7 @@
 			bind:value={$character.Class}
 			name="Class"
 			id="Class"
-			style="font-family: fantasy; font-size: 20px;"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
 		>
 			<option value="Barbarian" class="option">Barbarian</option>
 			<option value="Bard" class="option">Bard</option>
@@ -92,17 +119,6 @@
 			<option value="Wizard" class="option">Wizard</option>
 		</select>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="Background">Background:</label>
-		<select
-			bind:value={$character.Background}
-			name="Background"
-			id="Background"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option value="Acolyte" class="option">Acolyte</option>
-		</select>
-	</div>
 	<div class="createOptions" style="padding-left: 10px; padding-right: 5px;">
 		<label class="labelPg1" for="hitpoints" style="padding-right: 5px;">Rolled Hitpoints:</label>
 		<input
@@ -110,7 +126,7 @@
 			type="number"
 			id="hitpoints"
 			name="hitpoints"
-			style="font-family: fantasy; font-size: 20px;"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
 		/>
 	</div>
 	<div class="createOptions">
@@ -120,7 +136,7 @@
 			type="text"
 			id="name"
 			name="name"
-			style="font-family: fantasy; font-size: 20px;"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
 		/>
 	</div>
 </div>
