+++ my-app/src/lib/Fourth.svelte
@@ -1,178 +1,243 @@
 <script>
-	import Mylistitem from '$lib/Mylistitem.svelte';
-	import { onMount } from 'svelte';
-	import { writable } from 'svelte/store';
 	export let character;
-
-	/**
-	 * @type {HTMLInputElement}
-	 */
-	let search;
-
-	/**
-	 * @type {HTMLUListElement}
-	 */
-	let allFeats;
-
-	/**
-	 * @type {HTMLUListElement}
-	 */
-	let selectedFeats;
-
-	let featsArray = writable([]);
-
-	/**
-	 * Load feats from the API and populate the lists.
-	 */
-	async function loadFeats() {
-		try {
-			/**
-			 * @type {any}
-			 */
-			let data = { next: 'https://api.open5e.com/v2/feats/' };
-			do {
-				let resp = await fetch(data.next);
-				data = await resp.json();
-				data.results = data.results.map((/** @type {any} */ result) => {
-					return {
-						...result,
-						isSelected: false,
-						isShown: true
-					};
-				});
-				$featsArray = [...$featsArray, ...data.results];
-			} while (data.next);
-		} catch (error) {
-			console.error('Error loading feats:', error);
-		}
-	}
-
-	function onChangeFeat() {
-		$character.chosenFeats = getSelectedFeats().map((feat) => ({
-			name: feat.name,
-			description: feat.desc
-		}));
-	}
-
-	function getSelectedFeats() {
-		return $featsArray.filter((feat) => {
-			return feat.isSelected;
-		});
-	}
-
-	/**
-	 * Filter feats based on the search text.
-	 * @param {string} searchText
-	 */
-	function filterFeats(searchText) {
-		const lowerCaseText = searchText.toLowerCase();
-		$featsArray = $featsArray.map((feat) => {
-			function matches() {
-				const nameMatch = feat.name.toLowerCase().includes(lowerCaseText);
-				const descMatch = feat.desc?.toLowerCase().includes(lowerCaseText);
-				return nameMatch || descMatch;
-			}
-			feat.isShown = matches();
-			return feat;
-		});
-	}
-
-	onMount(() => {
-		loadFeats();
-
-		search.addEventListener('input', () => {
-			filterFeats(search.value);
-		});
-	});
+	const onStrengthChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Strength = event.target.value;
+	};
+	const onDexterityChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Dexterity = event.target.value;
+	};
+	const onConstitutionChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Constitution = event.target.value;
+	};
+	const onIntelligenceChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Intelligence = event.target.value;
+	};
+	const onWisdomChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Wisdom = event.target.value;
+	};
+	const onCharismaChange = (/** @type {{ target: { value: any; }; }} */ event) => {
+		$character.Charisma = event.target.value;
+	};
 </script>
 
-<div style="display: flex; flex-direction: row; text-align: center;">
-	<div class="container">
-		<div class="scrollable-container" style="margin-top: 20px">
-			<div
-				style="background-color: var(--button-bg); display: flex; flex-direction: column; justify-content: center; "
-			>
-				<h1>Selected Feats</h1>
-				<div style="text-align: center; justify-content: center; margin: 10px;">
-					<ul bind:this={selectedFeats} id="selected">
-						{#each $featsArray as feat}
-							{#if feat.isSelected}
-								<li>
-									<Mylistitem
-										name={feat.name}
-										desc={feat.desc}
-										isSelected={feat.isSelected}
-										on:selectToggled={(ev) => {
-											feat.isSelected = ev.detail.isSelected;
-											onChangeFeat();
-										}}
-									></Mylistitem>
-								</li>
-							{/if}
-						{/each}
-					</ul>
-				</div>
-			</div>
-		</div>
+<div class="wrapper">
+	<div class="createOptions">
+		<label class="labelPg1" for="Strength">Strength:</label>
+		<select
+			bind:value={$character.Strength}
+			name="Strength"
+			id="Strength"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
 	</div>
-	<div class="container">
-		<div class="scrollable-container">
-			<div style="justify-content: center; background-color: var(--button-bg); align-items:center ">
-				<h1>All Feats</h1>
-				<input
-					bind:this={search}
-					type="search"
-					placeholder="Search..."
-					name="feat"
-					id="searchField"
-				/>
-				<div style="text-align: center; justify-content: center; margin: 10px;">
-					<ul bind:this={allFeats} id="resultList">
-						{#each $featsArray as feat}
-							{#if !feat.isSelected && feat.isShown}
-								<li>
-									<Mylistitem
-										name={feat.name}
-										desc={feat.desc}
-										isSelected={feat.isSelected}
-										on:selectToggled={(ev) => {
-											feat.isSelected = ev.detail.isSelected;
-											onChangeFeat();
-										}}
-									></Mylistitem>
-								</li>
-							{/if}
-						{/each}
-					</ul>
-				</div>
-			</div>
-		</div>
+	<div class="createOptions">
+		<label class="labelPg1" for="dexterity">Dexterity:</label>
+		<select
+			bind:value={$character.Dexterity}
+			name="dexterity"
+			id="dexterity"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
+	</div>
+	<div class="createOptions">
+		<label class="labelPg1" for="constitution">Constitution:</label>
+		<select
+			bind:value={$character.Constitution}
+			name="constitution"
+			id="constitution"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
+	</div>
+	<div class="createOptions">
+		<label class="labelPg1" for="intelligence">Intelligence:</label>
+		<select
+			bind:value={$character.Intelligence}
+			name="intelligence"
+			id="intelligence"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
+	</div>
+	<div class="createOptions">
+		<label class="labelPg1" for="wisdom">Wisdom:</label>
+		<select
+			bind:value={$character.Wisdom}
+			name="wisdom"
+			id="wisdom"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
+	</div>
+	<div class="createOptions">
+		<label class="labelPg1" for="charisma">Charisma:</label>
+		<select
+			bind:value={$character.Charisma}
+			name="charisma"
+			id="charisma"
+			style="font-family: 'Franklin Gothic Medium', 'Arial Narrow', Arial, sans-serif; font-size: 20px;"
+		>
+			<option class="option" value={1}>1</option>
+			<option class="option" value={2}>2</option>
+			<option class="option" value={3}>3</option>
+			<option class="option" value={4}>4</option>
+			<option class="option" value={5}>5</option>
+			<option class="option" value={6}>6</option>
+			<option class="option" value={7}>7</option>
+			<option class="option" value={8}>8</option>
+			<option class="option" value={9}>9</option>
+			<option class="option" value={10}>10</option>
+			<option class="option" value={11}>11</option>
+			<option class="option" value={12}>12</option>
+			<option class="option" value={13}>13</option>
+			<option class="option" value={14}>14</option>
+			<option class="option" value={15}>15</option>
+			<option class="option" value={16}>16</option>
+			<option class="option" value={17}>17</option>
+			<option class="option" value={18}>18</option>
+			<option class="option" value={19}>19</option>
+			<option class="option" value={20}>20</option>
+		</select>
 	</div>
 </div>
 
 <style>
-	.scrollable-container {
-		height: 900px;
-		overflow-y: auto;
-		padding: 10px;
-		box-sizing: border-box;
-	}
-
-	.container {
-		width: 50%;
-		padding: 10px;
-		box-sizing: border-box;
-	}
-
-	#selected {
-		list-style: none;
-		padding: 0;
+	.wrapper {
+		display: flex;
+		flex-wrap: wrap;
+		background-color: #d9d9d9;
+		width: 100vw;
+		height: 100%;
+		justify-content: center;
+		align-items: center;
+		gap: 20px;
 		margin: 0;
+		flex-direction: row;
 	}
 
-	#resultList {
-		list-style: none;
+	.createOptions {
+		display: flex;
+		flex: wrap;
+		justify-content: center;
+		align-items: center;
+		min-width: 20em;
+		min-height: 5em;
+		max-width: 20em;
+		max-height: 5em;
+		background-color: var(--button-bg);
+		margin: 60px;
 		padding: 0;
-		margin: 0;
+		border: solid black 0.5px;
+	}
+
+	.labelPg1 {
+		font-size: 20px;
+		padding-right: 10px;
+	}
+
+	.option {
+		font-size: 20px;
 	}
 </style>
