+++ my-app/src/lib/Third.svelte
@@ -1,205 +1,146 @@
 <script>
+	// @ts-ignore
+	import { json } from '@sveltejs/kit';
 	export let character;
-	const onStrengthChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Strength = event.target.value;
-	};
-	const onDexterityChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Dexterity = event.target.value;
-	};
-	const onConstitutionChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Constitution = event.target.value;
-	};
-	const onIntelligenceChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Intelligence = event.target.value;
-	};
-	const onWisdomChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Wisdom = event.target.value;
-	};
-	const onCharismaChange = (/** @type {{ target: { value: any; }; }} */ event) => {
-		$character.Charisma = event.target.value;
+
+	const onSkillChanged = (
+		/** @type {Event & { currentTarget: EventTarget & HTMLInputElement; }} */ event
+	) => {
+		// @ts-ignore
+		const skillId = event.target.id;
+		// @ts-ignore
+		const isChecked = event.target.checked;
+		// @ts-ignore
+		console.log(skillId, isChecked);
+		// @ts-ignore
+		if (isChecked) {
+			// @ts-ignore
+			if (!$character.ProficientSkills.includes(skillId)) {
+				// @ts-ignore
+				$character.ProficientSkills.push(skillId);
+			}
+		} else {
+			// @ts-ignore
+			if ($character.ProficientSkills.includes(skillId)) {
+				// @ts-ignore
+				$character.ProficientSkills = $character.ProficientSkills.filter(
+					(/** @type {any} */ skill) => skill !== skillId
+				);
+			}
+		}
+		// @ts-ignore
+		console.log($character);
 	};
 </script>
 
 <div class="wrapper">
-	<div class="createOptions">
-		<label class="labelPg1" for="Strength">Strength:</label>
-		<select
-			bind:value={$character.Strength}
-			name="Strength"
-			id="Strength"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="athletics"
+			>Athletics
+			<input on:input={(e) => onSkillChanged(e)} id="athletics" type="checkbox" />
+		</label>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="dexterity">Dexterity:</label>
-		<select
-			bind:value={$character.Dexterity}
-			name="dexterity"
-			id="dexterity"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="acrobatics"
+			>Acrobatics
+			<input on:input={(e) => onSkillChanged(e)} id="acrobatics" type="checkbox" />
+		</label>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="constitution">Constitution:</label>
-		<select
-			bind:value={$character.Constitution}
-			name="constitution"
-			id="constitution"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="sleight-of-hand"
+			>Sleight of Hand
+			<input on:input={(e) => onSkillChanged(e)} id="sleight-of-hand" type="checkbox" />
+		</label>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="intelligence">Intelligence:</label>
-		<select
-			bind:value={$character.Intelligence}
-			name="intelligence"
-			id="intelligence"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="stealth"
+			>Stealth
+			<input on:input={(e) => onSkillChanged(e)} id="stealth" type="checkbox" />
+		</label>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="wisdom">Wisdom:</label>
-		<select
-			bind:value={$character.Wisdom}
-			name="wisdom"
-			id="wisdom"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="arcana"
+			>Arcana
+			<input on:input={(e) => onSkillChanged(e)} id="arcana" type="checkbox" />
+		</label>
 	</div>
-	<div class="createOptions">
-		<label class="labelPg1" for="charisma">Charisma:</label>
-		<select
-			bind:value={$character.Charisma}
-			name="charisma"
-			id="charisma"
-			style="font-family: fantasy; font-size: 20px;"
-		>
-			<option class="option" value={1}>1</option>
-			<option class="option" value={2}>2</option>
-			<option class="option" value={3}>3</option>
-			<option class="option" value={4}>4</option>
-			<option class="option" value={5}>5</option>
-			<option class="option" value={6}>6</option>
-			<option class="option" value={7}>7</option>
-			<option class="option" value={8}>8</option>
-			<option class="option" value={9}>9</option>
-			<option class="option" value={10}>10</option>
-			<option class="option" value={11}>11</option>
-			<option class="option" value={12}>12</option>
-			<option class="option" value={13}>13</option>
-			<option class="option" value={14}>14</option>
-			<option class="option" value={15}>15</option>
-			<option class="option" value={16}>16</option>
-			<option class="option" value={17}>17</option>
-			<option class="option" value={18}>18</option>
-			<option class="option" value={19}>19</option>
-			<option class="option" value={20}>20</option>
-		</select>
+	<div class="skill">
+		<label class="label" for="history"
+			>History
+			<input on:input={(e) => onSkillChanged(e)} id="history" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="investigation"
+			>Investigation
+			<input on:input={(e) => onSkillChanged(e)} id="investigation" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="nature"
+			>Nature
+			<input on:input={(e) => onSkillChanged(e)} id="nature" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="religion"
+			>Religion
+			<input on:input={(e) => onSkillChanged(e)} id="religion" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="animal-handling"
+			>Animal Handling
+			<input on:input={(e) => onSkillChanged(e)} id="animal-handling" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="insight"
+			>Insight
+			<input on:input={(e) => onSkillChanged(e)} id="insight" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="medicine"
+			>Medicine
+			<input on:input={(e) => onSkillChanged(e)} id="medicine" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="perception"
+			>Perception
+			<input on:input={(e) => onSkillChanged(e)} id="perception" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="survival"
+			>Survival
+			<input on:input={(e) => onSkillChanged(e)} id="survival" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="deception"
+			>Deception
+			<input on:input={(e) => onSkillChanged(e)} id="deception" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="intimidation"
+			>Intimidation
+			<input on:input={(e) => onSkillChanged(e)} id="intimidation" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="performance"
+			>Performance
+			<input on:input={(e) => onSkillChanged(e)} id="performance" type="checkbox" />
+		</label>
+	</div>
+	<div class="skill">
+		<label class="label" for="persuasion"
+			>Persuasion
+			<input on:input={(e) => onSkillChanged(e)} id="persuasion" type="checkbox" />
+		</label>
 	</div>
 </div>
 
@@ -217,7 +158,7 @@
 		flex-direction: row;
 	}
 
-	.createOptions {
+	.skill {
 		display: flex;
 		flex: wrap;
 		justify-content: center;
@@ -232,12 +173,7 @@
 		border: solid black 0.5px;
 	}
 
-	.labelPg1 {
-		font-size: 20px;
-		padding-right: 10px;
-	}
-
-	.option {
+	.label {
 		font-size: 20px;
 	}
 </style>
