+++ my-app/src/routes/+layout.svelte
@@ -20,7 +20,7 @@
 }
     .header {
         background-color: var(--button-selected);
-        width: 100vw;
+        width: 100%;
         height: 150px;
         display: flex;
         justify-content: center;
@@ -28,12 +28,12 @@
         padding: 0;
         margin: 0;
         font-size: 60px;
-        border: solid black 0.5px;
+        border: solid black 1px;
         border-bottom: 0.25px;
     }
 
     .menu {
-        width: 100vw;
+        width: 100%;
         height: 100px;
         display: flex;
         background-color: var(--button-bg);
@@ -47,7 +47,7 @@
         justify-content: center;
         align-items: center;
         height: 100%;
-        width: 50vw;
+        width: 50%;
         font-size: 40px;
         border-bottom: 0.25px;
     }
